pyproject.toml
[project]
name = "Video-Analysis"
version = "0.1.0"
dependencies = [
    "langchain-google-genai>=0.3.0",
    "opencv-python>=4.10.0",
    "pydantic>=2.11.0",
    "python-dotenv>=1.0.0"
]